pyproject.toml
[project]
name = "acmbot"
version = "0.1.0"
description = "Default template for PDM package"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "苏向夜", email = "fu050409@163.com"},
]
dependencies = [
    "nonebot2[fastapi]>=2.3.2",
    "nonebot-adapter-onebot>=2.4.4",
    "nonebot-adapter-console>=0.6.0",
    "tomlkit>=0.13.0",
]
license = {text = "AGPLv3"}

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" },
    { name = "Console", module_name = "nonebot.adapters.console" }
]
plugins = []
plugin_dirs = ["src/plugins"]
builtin_plugins = ["echo"]

[tool.pdm]
distribution = false
